+++ cutils/src/main/java/com/carlos/cutils/util/AccessibilityServiceUtils.kt
@@ -7,14 +7,16 @@ import android.view.accessibility.AccessibilityNodeInfo
  */
 object AccessibilityServiceUtils {
 
-    fun findAndClickFirstOneByText(text: String, accessibilityNodeInfo: AccessibilityNodeInfo) {
+    fun findAndClickFirstOneByText(text: String, accessibilityNodeInfo: AccessibilityNodeInfo?) {
+        if (accessibilityNodeInfo == null) return
         val node = accessibilityNodeInfo.findAccessibilityNodeInfosByText(text)
         if (node.isNotEmpty()) {
             node[0].performAction(AccessibilityNodeInfo.ACTION_CLICK)
         }
     }
 
-    fun findAndClickFirstOneById(resId: String, accessibilityNodeInfo: AccessibilityNodeInfo) {
+    fun findAndClickFirstOneById(resId: String, accessibilityNodeInfo: AccessibilityNodeInfo?) {
+        if (accessibilityNodeInfo == null) return
         val node = accessibilityNodeInfo.findAccessibilityNodeInfosByViewId(resId)
         if (node.isNotEmpty()) {
             node[0].performAction(AccessibilityNodeInfo.ACTION_CLICK)